TST: Moving DatetimeIndex related tests from test_timeseries.py and flake8 fixes
