+++ pandas/tests/indexes/datetimes/test_ops.py
@@ -5,8 +5,8 @@ import pandas as pd
 import pandas.tslib as tslib
 import pandas.util.testing as tm
 from pandas import (DatetimeIndex, PeriodIndex, Series, Timestamp, Timedelta,
-    date_range, TimedeltaIndex, _np_version_under1p10, Index, datetime,
-    Float64Index)
+                    date_range, TimedeltaIndex, _np_version_under1p10, Index,
+                    datetime, Float64Index)
 
 from pandas.tests.test_base import Ops
