+++ setup.py
@@ -640,6 +640,7 @@ setup(name=DISTNAME,
                 'pandas.tests',
                 'pandas.tests.frame',
                 'pandas.tests.indexes',
+                'pandas.tests.indexes.datetimes',
                 'pandas.tests.groupby',
                 'pandas.tests.series',
                 'pandas.tests.formats',
